pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "corpus-callosum"
version = "0.5.0"
description = "Unified learning and knowledge management toolkit with RAG, flashcards, and video transcription"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "chromadb>=1.0.20",
  "sentence-transformers>=3.0.0",
  "rank-bm25>=0.2.2",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "pypdf>=5.1.0",
  "python-docx>=1.1.0",
  "beautifulsoup4>=4.12.0",
  "markdownify>=0.12.0",
  "striprtf>=0.0.26",
  "click>=8.1.0",
  "ollama>=0.4.0",
  "mcp[cli]>=1.20.0",
  # Security dependencies
  "keyring>=25.0.0",
  "cryptography>=41.0.0",
  "typer[all]>=0.12.0",
  "rich>=13.0.0",
  "structlog>=24.0.0",
  "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.4.0",
  "mypy>=1.10.0",
  "types-PyYAML>=6.0.0",
]
observability = [
  "opentelemetry-api>=1.20.0",
  "opentelemetry-sdk>=1.20.0",
  "opentelemetry-exporter-otlp>=1.20.0",
  "opentelemetry-instrumentation-fastapi>=0.41b0",
  "opentelemetry-instrumentation-httpx>=0.41b0",
  "openllmetry>=0.3.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
# MCP Server
corpus-mcp-server = "mcp_server.server:main"

# Individual Tool CLIs
corpus-rag = "tools.rag.cli:main"
corpus-flashcards = "tools.flashcards.cli:main"
corpus-summaries = "tools.summaries.cli:main"
corpus-quizzes = "tools.quizzes.cli:main"
corpus-video = "tools.video.cli:main"

# Orchestrations
corpus-orchestrate = "orchestrations.cli:main"

# Database management
corpus-db = "db.management:main"

# Security and configuration tools
corpus-secrets = "utils.manage_secrets:main"
corpus-api-keys = "utils.manage_keys:main"

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E",      # pycodestyle errors
  "W",      # pycodestyle warnings
  "F",      # Pyflakes
  "I",      # isort
  "B",      # flake8-bugbear
  "C4",     # flake8-comprehensions
  "UP",     # pyupgrade
  "ARG",    # flake8-unused-arguments
  "SIM",    # flake8-simplify
  "PTH",    # flake8-use-pathlib
  "ERA",    # eradicate (commented-out code)
  "PL",     # Pylint
  "RUF",    # Ruff-specific rules
]
ignore = [
  "E501",    # line too long (handled by formatter)
  "PLR0913", # too many arguments
  "PLR2004", # magic value comparison
  "PLR0915", # too many statements
  "PLW0603", # global statement (valid for module-level singletons)
  "SIM108",  # use ternary operator (not always more readable)
  "PLC0415", # import not at top level (intentional for optional deps)
]

[tool.ruff.lint.per-file-ignores]
"tests/test_smoke.py" = ["E402"]  # imports after sys.path manipulation

[tool.ruff.lint.isort]
known-first-party = ["config", "db", "llm", "mcp_server", "orchestrations", "tools", "utils"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
  "chromadb.*",
  "sentence_transformers.*",
  "rank_bm25.*",
  "httpx.*",
  "opentelemetry.*",
  "openllmetry.*",
  "docx.*",
  "bs4.*",
  "markdownify.*",
  "striprtf.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
  "ignore::DeprecationWarning",
]
